+++ components/modal/basic/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.16.0"></a>
+# 1.16.0 (2019-09-19)
+
+
+### Bug Fixes
+
+* classList.toggle second argument not supported by IE 11 ([8709ed6](https://github.com/SUI-Components/schibsted-spain-components/commit/8709ed6))
+
+
+
 <a name="1.15.0"></a>
 # 1.15.0 (2019-08-14)